+++ views/default/forms/privacy_notification/acceptance.php
@@ -36,22 +36,9 @@ echo elgg_view('privacy_notification/acceptance/extend', $vars);
 
 ?>
 
-
 <div class="elgg-foot">
 <?php
-
     echo elgg_view('input/hidden', array('name' => 'user_guid', 'value' => $user->getGUID()));
-
-    // if (PrivacyNotificationOptions::isAccountRemovalBtnEnabled()) {
-    //     echo elgg_view('output/url', array(
-    //         'id' => 'remove_account_btn',
-    //         'href' => elgg_normalize_url("account_removal/$user->username"),
-    //         'class' => 'elgg-button elgg-button-submit elgg-button-delete elgg-size-small',
-    //         'text' => elgg_echo('privacy_notification:btn:remove_account'),
-    //         'title' => elgg_echo('privacy_notification:btn:remove_account'),
-    //     ));
-    // }
-    
     $btn = elgg_view_field([
         '#type' => 'submit',
         'id' => 'privacy_terms_btn',