+++ views/default/resources/privacy_notification/index.php
@@ -7,22 +7,21 @@
 use PrivacyNotification\PrivacyNotificationOptions;
 
 if (!elgg_is_logged_in()) {
+    
     $user_guid = get_input('user_guid');
     $invitecode = get_input('invitecode');
 
-    $user = get_entity($user_guid);
+    $user = get_entity($user_guid?$user_guid:0);
     if (PrivacyNotificationOptions::privacyNotificationIsSet() && PrivacyNotificationOptions::hasAcceptPN($user)) {
-        forward(elgg_get_site_url());
+        pn_forward(elgg_get_site_url());
     }
 
-    if (elgg_validate_invite_code($user->username, $invitecode)) {
+    if ($user && elgg_validate_invite_code($user->username, $invitecode)) {
         $vars['user_guid'] = $user_guid;
         $content = elgg_view_form('privacy_notification/acceptance', [], $vars);
-        $page_type = 'content';
     } else {
         // just show the privacy notifications
         $content = PrivacyNotificationOptions::getPrivacyNotificationText();
-        $page_type = 'one_column';
     }
 } 
 else if (
@@ -30,21 +29,24 @@ else if (
         !PrivacyNotificationOptions::hasAcceptPN()) {
 
     $content = elgg_view_form('privacy_notification/acceptance', [], $vars);
-    $page_type = 'content';
 } 
 else {
-    forward(elgg_get_site_url());
+    pn_forward(elgg_get_site_url());
 }
 
+$simple_view = get_input('s');
 
-$params = array( 
+if (!$simple_view || !empty($simple_view)) {
+    echo $content;
+    return;
+}
+
+$params = [
     'title' => elgg_echo('privacy_notification:index:title'),
     'content' => $content,
-    // 'sidebar' => $sidebar,
     'filter' => false,
     'class' => 'elgg-river-layout',
-);
+];
 
-$body = elgg_view_layout($page_type, $params);
-
+$body = elgg_view_layout('privacy', $params);
 echo elgg_view_page("", $body);